+++ checker/bin-devel/test-daikon-part1.sh
@@ -8,7 +8,12 @@ echo "SHELLOPTS=${SHELLOPTS}"
 
 SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 # shellcheck disable=SC1090# In newer shellcheck than 0.6.0, pass: "-P SCRIPTDIR" (literally)
-source "$SCRIPTDIR"/build.sh
+export ORG_GRADLE_PROJECT_useJdk17Compiler=true
+source "$SCRIPTDIR"/clone-related.sh
+
+# Run assembleForJavac because it does not build the javadoc, so it is faster than assemble.
+echo "running \"./gradlew assembleForJavac\" for checker-framework"
+./gradlew assembleForJavac --console=plain -Dorg.gradle.internal.http.socketTimeout=60000 -Dorg.gradle.internal.http.connectionTimeout=60000
 
 
 # daikon-typecheck: 15 minutes