+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/node/AbstractNodeVisitor.java
@@ -384,4 +384,9 @@ public abstract class AbstractNodeVisitor<R, P> implements NodeVisitor<R, P> {
   public R visitExpressionStatement(ExpressionStatementNode n, P p) {
     return visitNode(n, p);
   }
+
+  @Override
+  public R visitDeconstructorPattern(DeconstructorPatternNode n, P p) {
+    return visitNode(n, p);
+  }
 }