+++ framework/src/main/java/org/checkerframework/framework/ajava/JointJavacJavaParserVisitor.java
@@ -161,6 +161,11 @@ import java.util.Optional;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.TreeUtils;
+import org.checkerframework.javacutil.TreeUtilsAfterJava11.BindingPatternUtils;
+import org.checkerframework.javacutil.TreeUtilsAfterJava11.CaseUtils;
+import org.checkerframework.javacutil.TreeUtilsAfterJava11.InstanceOfUtils;
+import org.checkerframework.javacutil.TreeUtilsAfterJava11.SwitchExpressionUtils;
+import org.checkerframework.javacutil.TreeUtilsAfterJava11.YieldUtils;
 
 /**
  * A visitor that processes javac trees and JavaParser nodes simultaneously, matching corresponding
@@ -281,7 +286,7 @@ public abstract class JointJavacJavaParserVisitor extends SimpleTreeVisitor<Void
   public Void visitBindingPattern17(Tree javacTree, Node javaParserNode) {
     PatternExpr patternExpr = castNode(PatternExpr.class, javaParserNode, javacTree);
     processBindingPattern(javacTree, patternExpr);
-    VariableTree variableTree = TreeUtils.bindingPatternTreeGetVariable(javacTree);
+    VariableTree variableTree = BindingPatternUtils.getVariable(javacTree);
     // The name expression can be null, even when a name exists.
     if (variableTree.getNameExpression() != null) {
       variableTree.getNameExpression().accept(this, patternExpr.getName());
@@ -418,8 +423,7 @@ public abstract class JointJavacJavaParserVisitor extends SimpleTreeVisitor<Void
     processCase(javacTree, node);
     // Java 12 introduced multiple label cases:
     List<Expression> labels = node.getLabels();
-    List<? extends ExpressionTree> treeExpressions =
-        org.checkerframework.javacutil.TreeUtils.caseTreeGetExpressions(javacTree);
+    List<? extends ExpressionTree> treeExpressions = CaseUtils.getExpressions(javacTree);
     assert node.getLabels().size() == treeExpressions.size()
         : String.format(
             "node.getLabels() = %s, treeExpressions = %s", node.getLabels(), treeExpressions);
@@ -427,7 +431,7 @@ public abstract class JointJavacJavaParserVisitor extends SimpleTreeVisitor<Void
       treeExpressions.get(i).accept(this, labels.get(i));
     }
     if (javacTree.getStatements() == null) {
-      Tree javacBody = TreeUtils.caseTreeGetBody(javacTree);
+      Tree javacBody = CaseUtils.getBody(javacTree);
       Statement nodeBody = node.getStatement(0);
       if (javacBody.getKind() == Tree.Kind.EXPRESSION_STATEMENT) {
         javacBody.accept(this, node.getStatement(0));
@@ -844,7 +848,7 @@ public abstract class JointJavacJavaParserVisitor extends SimpleTreeVisitor<Void
     processInstanceOf(javacTree, node);
     javacTree.getExpression().accept(this, node.getExpression());
     if (node.getPattern().isPresent()) {
-      Tree bindingPattern = TreeUtils.instanceOfTreeGetPattern(javacTree);
+      Tree bindingPattern = InstanceOfUtils.getPattern(javacTree);
       visitBindingPattern17(bindingPattern, node.getPattern().get());
     } else {
       javacTree.getType().accept(this, node.getType());
@@ -1285,11 +1289,10 @@ public abstract class JointJavacJavaParserVisitor extends SimpleTreeVisitor<Void
 
     // Switch expressions are always parenthesized in javac but never in JavaParser.
     ExpressionTree expression =
-        ((ParenthesizedTree) TreeUtils.switchExpressionTreeGetExpression(javacTree))
-            .getExpression();
+        ((ParenthesizedTree) SwitchExpressionUtils.getExpression(javacTree)).getExpression();
     expression.accept(this, node.getSelector());
 
-    visitLists(TreeUtils.switchExpressionTreeGetCases(javacTree), node.getEntries());
+    visitLists(SwitchExpressionUtils.getCases(javacTree), node.getEntries());
     return null;
   }
 
@@ -1515,7 +1518,7 @@ public abstract class JointJavacJavaParserVisitor extends SimpleTreeVisitor<Void
       YieldStmt yieldStmt = castNode(YieldStmt.class, node, tree);
       processYield(tree, yieldStmt);
 
-      TreeUtils.yieldTreeGetValue(tree).accept(this, yieldStmt.getExpression());
+      YieldUtils.getValue(tree).accept(this, yieldStmt.getExpression());
       return null;
     }
     // JavaParser does not parse yields correctly: